Create a CredentialSource struct and update structs that rely on this data structure. This allows the CredentialSource source struct to be used on its own.
This is particularly useful for testing, where the SourceConfigGet is created from scratch (not using the encoding/json packages to pull in a response from the bindplane api).

This change is meant to improve useability for programs that import the sdk package, and not bindplane cli specifically. With that said, this does not introduce any breaking changes. Any package that imports bpcli/bindplane/sdk can still use the effects structs as if they were not changed. We are simply giving the user the optional to initialize and use the CredentialSource on its own.